apps/HipocratesAi-BackEnd/python_services/clinical_llm/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
openai==1.99.9
python-dotenv==1.0.1
supabase==2.10.0
